Set a driver speed limit
Setting a speed limit on the controller values rather than limiting the actual drive speed.  This way we can still keep the driver in check, but allow for tuning PID controllers at full speed. Also added the humandriver boolean which we use on the Raspberry Pi to prevent PID from accumulating error while a human driver is actively driving.